G1: Add initial HTC G1 modem support
G1 plugin is based on generic_at, with a bunch of stuff dropped
and simplified.  We use AT+CFUN=1 for powering on rather than having
a configurable init string.  We also manually set the default state
during init (the G1 appears to start in mode V0 by default).  The
device (/dev/smd0) is hardcoded.