udev: Add support for IFX device detection
The IFX device detection is pretty static, but instead of using
a static configuration file it is important to know when the device
node is actually present. For this udev is perfect. Adding a simple
udev rule is all that it takes:

KERNEL=="ttyIFX[0-9]*", ENV{OFONO_DRIVER}="ifx"

With this rule for every TTY with the kernel name like ttyIFX0, a new
modem will be added and the IFX modem plugin driver requested for it.