TODO: Remove SOCKS5 proxy task
The setup of GRPS context for MMS and its proxy configuration is now
done with host routes and dedicated interfaces. The need for a SOCKS5
proxy is no longer there.

The on-demand context activation can be considered a security risk
and activation via D-Bus method calls is easier to control in a
secure fashion. So this is not needed either.