gisi: Redesign ISI message processing
This new modem API supercedes the old client and server APIs.  It
properly queues resource subscriptions and version queries, as well as
caches reachability status for a given resource.

Also, this is a complete redesign of ISI message handling, adding
another layer of abstraction to multiplexing operations to a given
modem resource.