isi/voicecall: avoid side-effects of release
If a call was held automatically when an another call was dialed or
answered, it gets retrieved when the other call is released unless
auto change features is explicitly disabled.

Drop cause_type parameter because it is never used.

Use cause code for busy when releasing incoming or waiting calls.