calypso: Fix issue with name presentation and voice calls
The Calypso voice call driver uses an ofono_call object that is on the
stack. It is not automatically initilized to zero and thus causes to
have a name presentation that is not valid UTF-8. This in return makes
D-Bus fail and disconnect from the bus.

To fix this ensure that the ofono_call object is properly set to zero
first before populating it with values.