rilmodem/gprs-context: re-factored code to make all context ids gints.
The rilmodem gprs-context code internally stored context ids
as guints, and used -1 to indicate IDLE/disconnected state.
As rild and the core ofono code only used gints, this led
to casting issues, and also was confused the code, as the
gprs_context_data struct also includes a status field which
more clearly be used to check state.