src: Avoid state regressions
When the SIM state is switched to ready, the modem was assumed to be in
PRE_SIM state, which is not always the case for some modems. This caused
the modem to go to OFFLINE, and then to ONLINE, but the modem was
already ONLINE, and this implied removing some already available atoms.

The root cause is that for some modems when the radio is off you cannot
access the SIM, so the SIM atom is created when they move to ONLINE.